OLED screen functional
set I2C to use port 3 pins 0 and 1.  set input channels to be 2[1-4] for ADC input.  2 are for voltage and 2 are for thermistor.  thermistors are 10 k with 5k bridge from power.  Voltage inputs are intended for 12 volts with a shunt resistor to load of 10k and a bridge resistor to ground of 2 k to keep a max of 20 V less than 3.3V input.  Oled is run with gfx adafruit library that was adapted.  Thermistor calculation component causes a memory overrun and the display stops.  ADC is set to 10 bit after testing 8 bit.  The method of transferring input ADC to voltage and temperature and then converting cleanly to the display for visual observation is the difficulty.